Revert "Feature/apc#31: Added unit tests"
