+++ .gitignore
@@ -5,7 +5,6 @@
 .vscode
 .idea
 .npm
-.swc
 
 # testing
 /coverage
