+++ package.json
@@ -37,7 +37,6 @@
   "devDependencies": {
     "@testing-library/jest-dom": "^6.4.2",
     "@testing-library/react": "^14.2.2",
-    "@types/jest": "^29.5.12",
     "@types/lodash.debounce": "^4.0.9",
     "@types/node": "^20",
     "@types/react": "^18",
@@ -47,7 +46,6 @@
     "eslint": "^8",
     "eslint-config-next": "14.2.1",
     "jest": "^29.7.0",
-    "jest-environment-jsdom": "^29.7.0",
     "next-router-mock": "^0.9.13",
     "react-swipeable-views": "^0.14.0",
     "typescript": "^5"
