+++ src/components/common/CardItemList/CardItemList.tsx
@@ -10,32 +10,37 @@ import Image from "next/image";
 import { Typography } from "@mui/material";
 import DialogItemList from "../DialogItemList/DialogItemList";
 import { useState } from "react";
-import { Product } from "@/types/meli.types";
+import MeliService from "@/service/meli.service";
 
 interface CardItemListProps {
-  item: Product;
+  item: any;
 }
 
 const CardItemList = ({ item }: CardItemListProps) => {
   const [open, setOpen] = useState(false);
+  const [currentItem, setCurrentItem] = useState<any>(null);
 
-  const handleOnClick = async () => setOpen(true);
-
+  const handleOnClick = async () => {
+    const { data } = await MeliService.searchByItemId(item.id);
+    console.log("🚀 ~ handleOnClick ~ data:", data)
+    setCurrentItem(data);
+    setOpen(true);
+  };
+  
   const handleOnClose = () => setOpen(false);
 
   return (
     <>
       <StyledCategoryCard onClick={handleOnClick}>
         <StyledFirstIcon>
-          {item.thumbnail && (
-            <Image
-              src={item.thumbnail}
-              width={90}
-              height={90}
-              alt={item.thumbnail_id}
-              style={{ borderRadius: 8 }}
-            />
-          )}
+          {/* <img src={item.thumbnail} alt={item.thumbnail_id} /> */}
+          <Image
+            src={item.thumbnail}
+            width={90}
+            height={90}
+            alt={item.thumbnail_id}
+            style={{ borderRadius: 8 }}
+          />
         </StyledFirstIcon>
         <div>
           <StyledInfo>
@@ -48,7 +53,7 @@ const CardItemList = ({ item }: CardItemListProps) => {
           </StyledButton>
         </StyledLastIcon>
       </StyledCategoryCard>
-      <DialogItemList onClose={handleOnClose} open={open} item={item} />
+      <DialogItemList onClose={handleOnClose} open={open} item={currentItem} />
     </>
   );
 };
