+++ src/components/layout/Login/Login.styled.tsx
@@ -20,5 +20,6 @@ export const StyledForm = styled.div`
     text-align: center;
     padding: 32px;
     border-radius: 16px;
+    border: 1px solid #0D3B66;
     background: white;
 `
