+++ src/service/http.service.ts
@@ -1,15 +1,13 @@
 import { showSnackbar } from "@/redux/slice/snackbarSlice";
 import { store } from "@/redux/store";
-import axios, { AxiosInstance } from "axios";
+import axios, { AxiosError, AxiosInstance } from "axios";
 
 export abstract class HTTPService {
   protected instance: AxiosInstance;
   protected readonly baseURL: string;
-  protected readonly withToken: boolean;
 
-  public constructor(baseURL: string, withToken: boolean = true) {
+  public constructor(baseURL: string) {
     this.baseURL = baseURL;
-    this.withToken = withToken;
     this.instance = axios.create({
       baseURL,
     });
@@ -20,10 +18,6 @@ export abstract class HTTPService {
 
   private initializeRequestInterceptor = () => {
     this.instance.interceptors.request.use((request) => {
-      if (this.withToken) {
-        const token = store.getState().auth.token;
-        request.headers.Authorization = `Bearer ${token}`;
-      }
       return request;
     });
   };
@@ -35,7 +29,7 @@ export abstract class HTTPService {
   };
 
   private handleError = async (error: any) => {
-    const e = { ...error };
+    const e = { ...error }
     if (error.message === "Network Error" && !error.response) {
       store.dispatch(
         showSnackbar({
@@ -45,32 +39,11 @@ export abstract class HTTPService {
       );
     } else {
       const { data } = error?.response;
-      store.dispatch(showSnackbar({ message: data.status, severity: "error" }));
+      store.dispatch(
+        showSnackbar({ message: data.status, severity: "error" })
+      );
     }
 
     return Promise.reject(error);
   };
-
-  // private handleError(catchError: any, method: string) {
-  //   const error = new Error(
-  //     catchError?.response?.data?.message ||
-  //       catchError?.message ||
-  //       "Ha ocurrido un error inesperado."
-  //   );
-
-  //   const errorCode =
-  //     catchError.cause?.status || catchError.response?.status || 500;
-
-  //   console.log(
-  //     `CMS Service: Error ${errorCode} - ${error.message} occurred while executing ${method}`
-  //   );
-
-  //   return {
-  //     data: [],
-  //     error: {
-  //       code: errorCode,
-  //       message: error.message,
-  //     },
-  //   };
-  // }
 }
