+++ src/types/login.types.ts
@@ -1,20 +1,17 @@
 export interface LoginData {
-  email: string;
-  password: string;
+    email: string;
+    password: string;
 }
 
 export interface LoginUserData {
-  user: UserData;
-  // token: string;
+    user: UserData;
+    // token: string;
 }
 
 export interface UserData {
-  name: string;
-  surname: string;
-  username: string;
-  email: string;
-  createdDate?: Date;
-}
-export interface SignUpData extends LoginData {
-  name: string;
-}
+    name: string;
+    surname: string;
+    username: string;
+    email: string;
+    createdDate?: Date;
+  }
